pipeline/requirements.txt
yfinance>=0.2.36
requests>=2.31.0
anthropic>=0.40.0
python-dotenv>=1.0.1
